Add files via upload
The main purpose of this notes is to describe what our team has done for the hackathon. Although we have not accomplished a fully quantum solution yet, we provided a classical approach, which is presented in the attached notebook.

First and foremost, we uploaded the dataset for the Ising model from the website. Then, we identified the order parameter given by the magnetization $M_Z$ in the $z$ direction. Before building and computing the cost function, we decided to look at the data. Following this, we plotted $M_Z$ versus the magnetic field $h$. Since the number of phases equals 2, indeed there is a transition between two phases. One possible cost function is the magnetic susceptibility, $\chi = (\partial M_Z / \partial h)_T$, defined at constant temperature $T$. Assuming constant $T$, we computed the cost function numerically. As we were not able to associate the cost function with the corresponding PennyLane circuit, we could not apply PyTorch as in the given example. This might be the future development of this project.

In conclusion, we made a good attempt to address the proposed problem, although we haven't produced a fully quantum solution.